fix sequence - me enpoint - update response phase3

+++ api/src/main/java/com/api/planning/dto/request/EstimateUserStoryRequest.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class EstimateUserStoryRequest {
   @NotNull(message = "{error.estimation.required}")
   @Positive(message = "{error.estimation.positive}")
-  private Integer estimation;
+  private Double estimation;
 }
